It's not too late
At least I hope so. I always felt like it is too late for anything. Life is short. The more I think about it, the more it scares me. Time is a mysterious thing. Maybe the fear comes from the unknown. But what humans do in the past thousands of years, is exploring the unknown. Maybe it's never too late. One must stand in the perspective of humanity, can he see the big picture. Life is short indeed, but we humans still have a long way to go.