Find passion or make one
I like movies, TV shows, comics, but just because they are interesting. I never had a passion for something. As a result, I don't have any specialty. From now on, I need to give up most of the hobbies and only focus on the best. This applies to every aspect of my life.