My team fell apart
There was no warning or notification. I knew it from the missing of a member. I feel so bad to be a part of another failed project. It's September already. I should find another job at the beginning of the next year. I'll do my best to prepare for it.