Habit is hard to quit
I began to watch entertainment videos since serval months ago. Many times I want to quit this habit, but it happened so naturally. Before I know it, it's already bedtime, and I want to sleep early, so there is no time for things on my schedule. I know "first things first" is a good habit. So the next thing to do is to form this habit.